Moved package sqlParser into frontend/calciteCompiler
